refactor(core): define the shipped locales once
The list was restated four times — the generator's `[code, language]` pairs, the checker's codes, the app's `SUPPORT_LOCALES`, and one hand-written wrapper per language in EE. Adding a locale meant finding all four and creating a fifth file, and nothing failed if you missed one: the app would simply never offer the language, or the checker would never check it.

`src/translations/languages.ts` is now the source, and everything derives from it. It sits in `src/` rather than beside the tooling because the app imports it at runtime; the Node-only scripts are free to import app data, but not the other way round.
